pyproject.toml
[tool.poetry]
name = "gdpr-anonymizer"
version = "0.4.0"
description = "GDPR-compliant text anonymization system - Iteration 4 (Hexagonal Architecture)"
authors = ["GDPR Anonymizer Team <team@gdpr-anonymizer.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "anonymization", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.0.0"
fastapi = "^0.100.0"
uvicorn = "^0.27.0"
pyyaml = "^6.0"
python-dotenv = "^1.0.0"
httpx = "^0.27.0"
ollama = {version = "^0.1.0", optional = true}
anthropic = {version = "^0.18.0", optional = true}
openai = {version = "^1.0.0", optional = true}

[tool.poetry.extras]
ollama = ["ollama"]
claude = ["anthropic"]
openai = ["openai"]

[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
